ALMS-130 Implemented cognito credentials refresh mechanism
